requirements.txt
# Web / ASGI
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
python-multipart==0.0.19

# Twilio
twilio==9.4.1

# AI providers
anthropic==0.42.0
deepgram-sdk==3.8.0

# Utils
pyyaml==6.0.2
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.3
python-Levenshtein==0.26.1
rapidfuzz==3.10.1

# Dev / Test
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
